conversion is now a formula

+++ Behavior/UnitConvert.cs
@@ -19,8 +19,8 @@ public class UnitConvert : Behavior
 
         if (theseMatches != null && theseMatches.Count > 0 && theseMatches[0].Groups != null && theseMatches[0].Groups.Count == 5)
         {
-            decimal asNumeric = 0;
-            if (decimal.TryParse(theseMatches[0].Groups[1].Value, out asNumeric))
+            double asNumeric = 0;
+            if (double.TryParse(theseMatches[0].Groups[1].Value, out asNumeric))
             {
                 Console.WriteLine("let's try and convert...");
                 Behaver.Instance.SendMessage(message.Channel.Id, Conversion.Converter.Convert(asNumeric, theseMatches[0].Groups[2].Value, theseMatches[0].Groups[4].Value.ToLower()));
